+++ FinancialModelingPrepApi/Model/Calendars/DividendCalendarResponse.cs
@@ -11,13 +11,13 @@ namespace MatthiWare.FinancialModelingPrep.Model.Calendars
         public string Label { get; set; }
 
         [JsonPropertyName("adjDividend")]
-        public double AdjDividend { get; set; }
+        public decimal AdjDividend { get; set; }
 
         [JsonPropertyName("symbol")]
         public string Symbol { get; set; }
 
         [JsonPropertyName("dividend")]
-        public double? Dividend { get; set; }
+        public decimal? Dividend { get; set; }
 
         [JsonPropertyName("recordDate")]
         public string RecordDate { get; set; }
