+++ FinancialModelingPrepApi/Model/Calendars/IPOCalendarResponse.cs
@@ -26,6 +26,6 @@ namespace MatthiWare.FinancialModelingPrep.Model.Calendars
         public string PriceRange { get; set; }
 
         [JsonPropertyName("marketCap")]
-        public double? MarketCap { get; set; }
+        public decimal? MarketCap { get; set; }
     }
 }
