+++ FinancialModelingPrepApi/Model/CompanyValuation/QuoteResponse.cs
@@ -12,34 +12,34 @@ namespace MatthiWare.FinancialModelingPrep.Model.CompanyValuation
         public string Name { get; set; }
 
         [JsonPropertyName("price")]
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
 
         [JsonPropertyName("changesPercentage")]
-        public double? ChangesPercentage { get; set; }
+        public decimal? ChangesPercentage { get; set; }
 
         [JsonPropertyName("change")]
-        public double? Change { get; set; }
+        public decimal? Change { get; set; }
 
         [JsonPropertyName("dayLow")]
-        public double? DayLow { get; set; }
+        public decimal? DayLow { get; set; }
 
         [JsonPropertyName("dayHigh")]
-        public double? DayHigh { get; set; }
+        public decimal? DayHigh { get; set; }
 
         [JsonPropertyName("yearHigh")]
-        public double? YearHigh { get; set; }
+        public decimal? YearHigh { get; set; }
 
         [JsonPropertyName("yearLow")]
-        public double? YearLow { get; set; }
+        public decimal? YearLow { get; set; }
 
         [JsonPropertyName("marketCap")]
-        public double? MarketCap { get; set; }
+        public decimal? MarketCap { get; set; }
 
         [JsonPropertyName("priceAvg50")]
-        public double? PriceAvg50 { get; set; }
+        public decimal? PriceAvg50 { get; set; }
 
         [JsonPropertyName("priceAvg200")]
-        public double? PriceAvg200 { get; set; }
+        public decimal? PriceAvg200 { get; set; }
 
         [JsonPropertyName("volume")]
         public long? Volume { get; set; }
@@ -51,16 +51,16 @@ namespace MatthiWare.FinancialModelingPrep.Model.CompanyValuation
         public string Exchange { get; set; }
 
         [JsonPropertyName("open")]
-        public double? Open { get; set; }
+        public decimal? Open { get; set; }
 
         [JsonPropertyName("previousClose")]
-        public double? PreviousClose { get; set; }
+        public decimal? PreviousClose { get; set; }
 
         [JsonPropertyName("eps")]
-        public double? Eps { get; set; }
+        public decimal? Eps { get; set; }
 
         [JsonPropertyName("pe")]
-        public double? Pe { get; set; }
+        public decimal? Pe { get; set; }
 
         [JsonPropertyName("earningsAnnouncement")]
         public string EarningsAnnouncement { get; set; }
